+++ Sources/BartyCrouchTranslator/DeeplApi/Model/DeepLTranslateResponse.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+struct DeepLTranslateResponse: Decodable {
+    struct Translation: Decodable {
+        let detectedSourceLanguage: String
+        let text: String
+    }
+
+    let translations: [Translation]
+}